pyproject.toml
[project]
name = "butterfly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.2.1",
    "numpy==2.2.1",
    "open-clip-torch==2.30.0",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "scikit-learn>=1.6.1",
    "torch==2.3.1",
    "torchvision==0.18.1",
    "tyro==0.9.13",
    "gdown",
]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.isort]
line_length = 100
src_paths = ["."]

[tool.black]
line-length = 100